+++ app/src/androidTest/java/com/hedvig/app/feature/embark/screens/TextActionSetScreen.kt
@@ -6,7 +6,7 @@ import com.agoda.kakao.recycler.KRecyclerItem
 import com.agoda.kakao.recycler.KRecyclerView
 import com.agoda.kakao.text.KButton
 import com.hedvig.app.R
-import com.hedvig.app.feature.embark.passages.textactionset.TextActionFragment
+import com.hedvig.app.feature.embark.passages.textaction.TextActionFragment
 import com.kaspersky.kaspresso.screens.KScreen
 import org.hamcrest.Matcher
