+++ app/src/main/java/com/hedvig/app/feature/embark/passages/textaction/TextActionParameter.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.feature.embark.passages.textactionset
+package com.hedvig.app.feature.embark.passages.textaction
 
 import android.os.Parcelable
 import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
